requirements.txt
asyncpg==0.29.0
discord.py==2.3.2
jishaku==2.5.2
python-dotenv==1.0.1
tortoise-orm[asyncpg]==0.20.1
